change POST function and one send parameter

+++ app/services/user_service.py
@@ -1,5 +1,4 @@
 from fastapi import HTTPException
-
 from app.database import repository
 from app.database.database_connection import Collections
 from app.models.user import User
@@ -32,7 +31,7 @@ async def get_user_by_id(user_id: str):
         """
     try:
         user = await repository.get_by_id(Collections.users, user_id)
-        if not user:
+        if user is None:
             raise HTTPException(status_code=404, detail=f"User with ID {user_id} not found")
         return user
     except Exception as e:
@@ -50,6 +49,10 @@ async def add_user(new_user: User):
            ValueError: If the user object is null or the user ID already exists.
            Exception: If there is an error during the addition process.
        """
+    if new_user is None:
+        raise ValueError("User object is null")
+    if await get_user_by_id(new_user.id) is not None:
+        raise ValueError("User ID already exists")
     try:
         return await repository.add(Collections.users, new_user.dict())
     except Exception as e:
@@ -67,6 +70,8 @@ async def update_user(user_id: str, updated_data: User):
        Raises:
            ValueError: If the user object is null or the user is not found.
            Exception: If there is an error during the update process.
+           :param user_id:
+           :param updated_data:
        """
     try:
         return await repository.update(Collections.users, user_id, updated_data.dict())
@@ -85,6 +90,8 @@ async def delete_user(user_id: str):
          ValueError: If the user is not found.
          Exception: If there is an error during the deletion process.
      """
+    if await get_user_by_id(user_id) is None:
+        raise ValueError("User not found")
     try:
         return await repository.delete(Collections.users, user_id)
     except Exception as e:
